refactor: extract durationExpr
